requirements.txt
pymongo
discord.py
python-dotenv
requests
